i reap what i sow
i decided today i would do something cool... terrible idea. i wanted to do a cool image-manipulation component, i had
all these crazy ideas about sliders for dimensions and positioning and checkboxes to decide if you wanted to zoom or
cut the rest and picking opacity and including actually useful default options...
i put in two sliders and spent the rest of the day banging my head on the keyboard.
materialUI getting very mad that something is changing the component from UNCONTROLLED to CONTROLLED. but i thought
it was already controlled... it's probably because value is not set at the start. fuck me