Fixed sneaky input label bug
Bug: if you insert two images in a template, move to creator and open
both collapses, uploading an image on the second changes the first.
reason: because the input is display:none, the label needed its id to
forward events. because the id's were the same, the resolution found
the first one and called it

Now the player is marked as done when the he finishes the story

Co-authored-by: mettz <mattia.guazzaloca@gmail.com>
Co-authored-by: pollomarzo <pollo.marzo@gmail.com>
Co-authored-by: Muduard <muduard@gmail.com>